+++ test/src/tests/counter_test.dart
@@ -5,8 +5,10 @@ String _location = '========================== counter_test.dart';
 
 /// Testing the counter app
 Future<void> counterTest(WidgetTester tester) async {
+  // Retrieve the current count.
+  String start = CounterController().data;
   // Verify that our counter starts at 0.
-  expect(find.text('0'), findsOneWidget, reason: _location);
+  expect(find.text(start), findsOneWidget, reason: _location);
   expect(find.text('1'), findsNothing, reason: _location);
 
   // 9 counts
@@ -16,7 +18,10 @@ Future<void> counterTest(WidgetTester tester) async {
     await tester.pump();
   }
 
+  final end = int.parse(start) + 9;
+  start = end.toString();
+
   // Verify that our counter has incremented.
   expect(find.text('0'), findsNothing, reason: _location);
-  expect(find.text('9'), findsOneWidget, reason: _location);
+  expect(find.text(start), findsOneWidget, reason: _location);
 }